Ensure ember-cli knows the proper location of the addon entry point.
Ember CLI previously used `index.js` as a fallback path, but it should
have checked the `main` property first (which is being added in a pull
request). However, if `main` is used with `broccoli-asset-rev` we do not
get the right entry script (we would get `lib/asset-rev.js` instead of
`index.js`).

This fixes that issue by defining the custom `ember-addon.main` entry
point script.  This is backwards compatible.